Some basic changes, moved Interfaces to Utils folder
Moved function prototypes inside hooks class as they do not need to be global

Changed Utils::FindSignature return type to uintptr_t

Started to get rid of too many unneeded pointers. For now removed Interfaces class + pointer.

+++ Antario/Utils/Utils.h
@@ -8,7 +8,7 @@
 class Utils
 {
 public:
-    static uint64_t FindSignature(const char* szModule, const char* szSignature)
+    static uintptr_t FindSignature(const char* szModule, const char* szSignature)
     {
         const char* pat = szSignature;
         DWORD firstMatch = 0;
